+++ source/db_constraints/utils/package.d
@@ -2,6 +2,6 @@ module db_constraints.utils;
 
 public
 {
-    import db_constraints.utils.generickey;
+    import db_constraints.utils.meta;
     import db_constraints.utils.nullable;
 }